feat: add NearestNeighbourScaling

+++ src/main/java/ru/itmo/grafix/ui/models/SimpleScalingParams.java
@@ -4,7 +4,7 @@ import ru.itmo.grafix.core.scaling.Scaling;
 
 public class SimpleScalingParams extends ScalingParams{
 
-    public SimpleScalingParams(float width, float height, float biasX, float biasY, Scaling scalingMethod) {
+    public SimpleScalingParams(int width, int height, float biasX, float biasY, Scaling scalingMethod) {
         super(width, height, biasX, biasY, scalingMethod);
     }
 }
